91. Improving the Structure of Our Utility Scripts

+++ key-value-app/start-db.sh
@@ -1,35 +1,29 @@
-#!/bin/bash
-
-# MongoDB image and version
 MONGODB_IMAGE="mongodb/mongodb-community-server"
 MONGODB_TAG="7.0-ubuntu2204"
-
-# Container name
-CONTAINER_NAME="mongodb"
-
-
-
-# Key-value app credentials
-KEY_VALUE_DB="key-value-db"
-KEY_VALUE_USER="key-value-user"
-KEY_VALUE_PASSWORD="key-value-password"
-
+source .env.db
 
 # Root credentials
 ROOT_USER="root-user"
 ROOT_PASSWORD="root-password"
 
 # Connectivity
+source .env.network
 LOCALHOST_PORT=27017
 CONTAINER_PORT=27017
 
-NETWORK_NAME="keyvalue-net"
-
 # Storage
-VOLUME_NAME="keyvalue-data"
+source .env.volume
 VOLUME_CONTAINER_PATH="/data/db"
 
-# Start MongoDB container
+source setup.sh
+
+if [ "$(docker ps -q -f name=$DB_CONTAINER_NAME)" ]; then
+    echo "A container with the name $DB_CONTAINER_NAME already exists."
+    echo "The container will be removed when stopped."
+    echo "To stop the container, run: docker kill $DB_CONTAINER_NAME"
+    exit 1
+fi
+
 docker run --rm -d --name $DB_CONTAINER_NAME \
   -e MONGODB_INITDB_ROOT_USERNAME=$ROOT_USER \
   -e MONGODB_INITDB_ROOT_PASSWORD=$ROOT_PASSWORD \
